% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_predict_model}
\alias{fit_predict_model}
\title{Fit predictive model and return predictions}
\usage{
fit_predict_model(
  model_type,
  X_L,
  y_L,
  X_U,
  mtry = NULL,
  rf_engine = c("ranger", "randomForest"),
  rf_trees = 200,
  rf_min_node_size = 5,
  rf_num_threads = NULL,
  rf_seed = NULL
)
}
\arguments{
\item{model_type}{Type of model ("glm_correct", "glm_mis", "glm_wrong", "rf")}

\item{X_L}{Labeled covariates}

\item{y_L}{Labeled outcomes}

\item{X_U}{Unlabeled covariates}

\item{mtry}{RF mtry value}

\item{rf_engine}{Random forest engine ("ranger" or "randomForest")}

\item{rf_trees}{Number of trees}

\item{rf_min_node_size}{Minimum node size for ranger}

\item{rf_num_threads}{Threads (ranger)}

\item{rf_seed}{Random seed}
}
\value{
List containing model fit object and predictions on X_U.
}
\description{
Internal unified interface for GLMs and random forests.
}
\keyword{internal}
