% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_mean.R
\name{ppi_mean_rep_one}
\alias{ppi_mean_rep_one}
\title{One Monte Carlo replicate for PPI / PPI++}
\usage{
ppi_mean_rep_one(
  n,
  N,
  X_sampler_L,
  X_sampler_U = NULL,
  f_generator,
  eps_sampler = function(n) rnorm(n, 0, 1),
  model_type = c("glm_correct", "glm_mis", "glm_wrong", "rf"),
  ppi_type = c("naive", "PPI", "PPI++"),
  lambda_mode = c("plugin", "theory"),
  lambda_external = FALSE,
  n_external = ceiling(n/2),
  alpha = 0.05,
  delta = 0,
  seed = NULL,
  pred_noise = 0,
  PPI_crossfit = FALSE,
  PPIpp_crossfit = TRUE,
  PPIpp_independent_split = FALSE
)
}
\value{
A list with theta_hat, se, z, reject, lambda_hat, and variance components.
}
\description{
Internal engine for computing one repetition of (theta_hat, lambda_hat, SE, reject).
}
\keyword{internal}
