% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_ols_fit.R
\name{ppi_ols_fit}
\alias{ppi_ols_fit}
\title{Fit Prediction-Powered OLS (user-facing)}
\usage{
ppi_ols_fit(y, f_col, formula, data_l, data_u)
}
\arguments{
\item{y}{Name of outcome in labeled data (string or symbol)}

\item{f_col}{Name of column(s) containing predictions (in labeled/unlabeled)}

\item{formula}{RHS-only formula for covariates, e.g. ~ x1 + x2}

\item{data_l}{Labeled data.frame (must contain y, f_col, covariates)}

\item{data_u}{Unlabeled data.frame (must contain f_col, covariates)}
}
\value{
An object of class \code{"ppi_ols_fit"} with coef, vcov, se, n, N, call.
}
\description{
Fit Prediction-Powered OLS (user-facing)
}
