% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_predict_model_glm}
\alias{fit_predict_model_glm}
\title{Fit predictive model for GLM-based PPI/PPI++}
\usage{
fit_predict_model_glm(
  model_type,
  X_L,
  y_L,
  X_U,
  fold_index = NULL,
  mtry = NULL,
  rf_engine = c("ranger", "randomForest"),
  rf_trees = 200,
  rf_min_node_size = 5,
  rf_num_threads = NULL,
  f_generator = NULL
)
}
\arguments{
\item{model_type}{One of \code{"glm_correct"}, \code{"glm_mis"}, \code{"glm_wrong"}, \code{"rf"}, \code{"oracle"}.}

\item{X_L}{Labeled covariate matrix/data frame.}

\item{y_L}{Labeled response vector.}

\item{X_U}{Unlabeled covariates.}

\item{fold_index}{Optional vector of fold assignments for cross-fitting.}

\item{mtry}{Random forest mtry parameter.}

\item{rf_engine}{Random forest engine (\code{"ranger"} or \code{"randomForest"}).}

\item{rf_trees}{Number of trees for RF.}

\item{rf_min_node_size}{Minimum node size for RF.}

\item{rf_num_threads}{Number of threads for RF.}

\item{f_generator}{Required if \code{model_type = "oracle"}; function returning linear predictor η(x).}
}
\value{
A list with:
\item{fit}{The fitted model object.}
\item{fhat_L}{Predicted conditional means \eqn{\hat\mu_f(X_L)} on labeled data.}
\item{fhat_U}{Predicted conditional means \eqn{\hat\mu_f(X_U)} on unlabeled data.}
}
\description{
Fit predictive model for GLM-based PPI/PPI++
}
\keyword{internal}
