% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_n_ppi.R
\name{n_required_PP}
\alias{n_required_PP}
\title{Solve for required n (labeled sample size) for a desired power}
\usage{
n_required_PP(
  delta,
  var_f,
  var_res,
  N,
  alpha = 0.05,
  power = 0.8,
  warn_smallN = TRUE,
  smallN_threshold = 500,
  mode = c("error", "cap")
)
}
\arguments{
\item{delta}{Effect size (θ - θ_0)}

\item{var_f}{Variance of f(X), the predictor function}

\item{var_res}{Variance of (Y - f(X)), the residual}

\item{N}{Unlabeled sample size}

\item{alpha}{Significance level}

\item{power}{Target power (1 - β)}
}
\value{
Required labeled sample size n
}
\description{
Solve for required n (labeled sample size) for a desired power
}
\examples{
n_required_PP(delta = 0.1, var_f = 0.2, var_res = 0.5, N = 5000, alpha = 0.05, power = 0.8)
}
